refactor(sdk): unify getWriteBodyParams and adoptPublishedFolderState
Extract getWriteBodyParams, adoptPublishedFolderState, and hasRealWriteKey
into a shared packages/sdk/src/write-body-params.ts module, and re-point
both CipherBoxClient (client.ts, a thin delegating private method) and the
bin operations module (bin/index.ts, a direct import) at the single
implementation. Plan 04 already made the two getWriteBodyParams copies
byte-identical including the SC#2 fail-closed change, so this collapses
the drift surface without altering behavior. Also reconciles the resolve
path: client.ts previously went through the private resolvePublishedNode
helper while bin/index.ts inlined resolveIpnsRecord+fetchFromIpfs+JSON.parse
directly; the shared module standardizes on the inline form both already
proved behaviorally identical for this call site.
